feat: Implement real time publishing.
