+++ .idea/runConfigurations/go_test_internal.xml
@@ -1,0 +1,13 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="go test internal" type="GoTestRunConfiguration" factoryName="Go Test" nameIsGenerated="true">
+    <module name="plugin-pub-mssql" />
+    <working_directory value="$PROJECT_DIR$/internal" />
+    <parameters value="-ginkgo.progress" />
+    <framework value="gotest" />
+    <kind value="DIRECTORY" />
+    <package value="github.com/naveego/plugin-pub-mssql" />
+    <directory value="$PROJECT_DIR$/internal" />
+    <filePath value="$PROJECT_DIR$/" />
+    <method v="2" />
+  </configuration>
+</component>
