+++ version/version.go
@@ -4,13 +4,13 @@ import "github.com/coreos/go-semver/semver"
 
 var (
 	// VersionMajor is for an API incompatible changes.
-	VersionMajor int64 = 2
+	VersionMajor int64 = 3
 	// VersionMinor is for functionality in a backwards-compatible manner.
-	VersionMinor int64 = 1
+	VersionMinor int64 = 0
 	// VersionPatch is for backwards-compatible bug fixes.
-	VersionPatch int64 = 6
+	VersionPatch int64 = 0
 	// VersionPre indicates prerelease.
-	VersionPre string = "beta2"
+	VersionPre string = "beta1"
 	// VersionDev indicates development branch. Releases will be empty string.
 	VersionDev string
 )
